requirements.txt
arrow==0.6.0
diff-match-patch==20121119
Django==1.7.5
django-admin-tools==0.6.0
django-adminplus==0.3
django-import-export==0.2.8
docutils==0.12
feedparser==5.2.1
python-dateutil==2.4.2
pytz==2015.4
six==1.9.0
tablib==0.10.0
